pyproject.toml
[project]
name = "vit-mae"
version = "0.1.0"
description = "Vision Transformer Masked Autoencoder (ViT-MAE) implementation"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.21.0",
    "torch>=1.9.0",
    "Pillow>=8.0.0",
    "pyyaml>=6.0",
    "pandas>=1.3.0",
    "matplotlib>=3.10.7",
]

[tool.setuptools]
py-modules = ["vit_mae", "data_loader"]
packages = []

[tool.setuptools.package-data]
"*" = []

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

